Seprating interfaces from implementations
